+++ src/collective/solr/browser/maintenance.py
@@ -9,10 +9,6 @@ from plone.uuid.interfaces import IUUID, IUUIDAware
 from zope.interface import implementer
 from zope.component import queryUtility, queryAdapter
 
-try:
-    from collective.indexing.indexer import getOwnIndexMethod
-except ImportError:
-    getOwnIndexMethod = None
 from collective.solr.indexer import DefaultAdder
 from collective.solr.flare import PloneFlare
 from collective.solr.interfaces import ISolrConnectionManager
@@ -180,11 +176,6 @@ class SolrMaintenanceView(BrowserView):
 
         for path, obj in findObjects(self.context):
             if ICheckIndexable(obj)():
-                if getOwnIndexMethod:
-                    if getOwnIndexMethod(obj, "indexObject") is not None:
-                        log("skipping indexing of %r via private method.\n" % obj)
-                        continue
-
                 count += 1
                 if count <= skip:
                     continue