+++ CHANGES.rst
@@ -8,6 +8,16 @@ Changelog
 
 
 - Remove collective.js.showmore dependency
+- Remove plone.app.contenttypes conditionals in code
+  [timo]
+
+- Remove traces of collective.indexing
+  [timo]
+
+- Remove Plone 4.3 GS profile
+  [timo]
+
+- Remove AT conditionals
   [timo]
 
 9.0.0a1 (2021-11-19)